# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_thepantheon_api_internal_model.CapaVadeMecumCodigo:
    properties:
      Cabecalho:
        type: string
      grupo:
        type: string
      id:
        type: string
      nomecodigo:
        type: string
    type: object
...
  github_com_thepantheon_api_internal_model.CreateCapaVadeMecumCodigoRequest:
    properties:
      Cabecalho:
        type: string
      grupo:
        type: string
      nomecodigo:
        type: string
    required:
    - nomecodigo
    type: object
...
  github_com_thepantheon_api_internal_model.CreateVadeMecumCodigoRequest:
    properties:
      Cabecalho:
        type: string
      Normativo:
        type: string
      Ordem:
        type: string
      PARTE:
        type: string
      capitulo:
        type: string
      capitulotexto:
        type: string
      idcapitulo:
        type: string
      idcodigo:
        type: string
      idlivro:
        type: string
      idsecao:
        type: string
      idsubsecao:
        type: string
      idsubtitulo:
        type: string
      idtipo:
        type: string
      idtitulo:
        type: string
      livro:
        type: string
      livrotexto:
        type: string
      nomecodigo:
        type: string
      num_artigo:
        type: string
      secao:
        type: string
      secaotexto:
        type: string
      subsecao:
        type: string
      subsecaotexto:
        type: string
      subtitulo:
        type: string
      subtitulotexto:
        type: string
      tipo:
        type: string
      titulo:
        type: string
      titulotexto:
        type: string
    required:
    - nomecodigo
    type: object
...
  github_com_thepantheon_api_internal_model.UpdateCapaVadeMecumCodigoRequest:
    properties:
      grupo:
        type: string
    type: object
...
  github_com_thepantheon_api_internal_model.UpdateVadeMecumCodigoRequest:
    properties:
      Cabecalho:
        type: string
      Normativo:
        type: string
      Ordem:
        type: string
      PARTE:
        type: string
      capitulo:
        type: string
      capitulotexto:
        type: string
      idcapitulo:
        type: string
      idcodigo:
        type: string
      idlivro:
        type: string
      idsecao:
        type: string
      idsubsecao:
        type: string
      idsubtitulo:
        type: string
      idtipo:
        type: string
      idtitulo:
        type: string
      livro:
        type: string
      livrotexto:
        type: string
      nomecodigo:
        type: string
      num_artigo:
        type: string
      secao:
        type: string
      secaotexto:
        type: string
      subsecao:
        type: string
      subsecaotexto:
        type: string
      subtitulo:
        type: string
      subtitulotexto:
        type: string
      tipo:
        type: string
      titulo:
        type: string
      titulotexto:
        type: string
    type: object
...
  github_com_thepantheon_api_internal_model.VadeMecumCodigo:
    properties:
      Cabecalho:
        type: string
      Normativo:
        type: string
      Ordem:
        type: string
      PARTE:
        type: string
      capitulo:
        type: string
      capitulotexto:
        type: string
      created_at:
        type: string
      id:
        type: string
      idcapitulo:
        type: string
      idcodigo:
        type: string
      idlivro:
        type: string
      idsecao:
        type: string
      idsubsecao:
        type: string
      idsubtitulo:
        type: string
      idtipo:
        type: string
      idtitulo:
        type: string
      livro:
        type: string
      livrotexto:
        type: string
      nomecodigo:
        type: string
      num_artigo:
        type: string
      secao:
        type: string
      secaotexto:
        type: string
      subsecao:
        type: string
      subsecaotexto:
        type: string
      subtitulo:
        type: string
      subtitulotexto:
        type: string
      tipo:
        type: string
      titulo:
        type: string
      titulotexto:
        type: string
      updated_at:
        type: string
    type: object
  github_com_thepantheon_api_internal_model.VadeMecumCodigoGroup:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_thepantheon_api_internal_model.VadeMecumCodigo'
        type: array
      nomecodigo:
        type: string
      priority:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@pantheonconcursos.com.br
    name: API Support
  description: API REST para sistema de concursos com autenticação JWT
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: ThePantheon API
  version: "1.0"
paths:
...
  /vade-mecum/codigos:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_thepantheon_api_internal_model.VadeMecumCodigo'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar codigos
      tags:
      - vade-mecum-codigos
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados do codigo
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_thepantheon_api_internal_model.CreateVadeMecumCodigoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_thepantheon_api_internal_model.VadeMecumCodigo'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Criar item na categoria codigos
      tags:
      - vade-mecum-codigos
  /vade-mecum/codigos/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remover codigo
      tags:
      - vade-mecum-codigos
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thepantheon_api_internal_model.VadeMecumCodigo'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obter codigo por ID
      tags:
      - vade-mecum-codigos
    put:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: Dados
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_thepantheon_api_internal_model.UpdateVadeMecumCodigoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thepantheon_api_internal_model.VadeMecumCodigo'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualizar codigo
      tags:
      - vade-mecum-codigos
  /vade-mecum/codigos/capas:
    get:
      parameters:
      - description: Filtro por nomecodigo
        in: query
        name: nomecodigo
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_thepantheon_api_internal_model.CapaVadeMecumCodigo'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar capas de códigos ou buscar por nome específico
      tags:
      - vade-mecum-codigos
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados da capa
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_thepantheon_api_internal_model.CreateCapaVadeMecumCodigoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_thepantheon_api_internal_model.CapaVadeMecumCodigo'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Criar capa de vade-mécum código
      tags:
      - vade-mecum-codigos
  /vade-mecum/codigos/capas/{id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: Identificador da capa
        in: path
        name: id
        required: true
        type: string
      - description: Campos para atualização
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_thepantheon_api_internal_model.UpdateCapaVadeMecumCodigoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thepantheon_api_internal_model.CapaVadeMecumCodigo'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualizar capa de vade-mécum código
      tags:
      - vade-mecum-codigos
  /vade-mecum/codigos/grouped:
    get:
      parameters:
      - collectionFormat: csv
        description: Prioridade dos grupos, pode ser separado por vírgula ou repetido
        in: query
        items:
          type: string
        name: priority
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_thepantheon_api_internal_model.VadeMecumCodigoGroup'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar codigos agrupados por nomecodigo
      tags:
      - vade-mecum-codigos
  /vade-mecum/codigos/import:
    post:
      consumes:
      - multipart/form-data
      description: Importa registros utilizando um arquivo Excel (.xlsx) com cabeçalho
        padrão
      parameters:
      - description: Arquivo Excel (.xlsx)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Importar codigos via Excel
      tags:
      - vade-mecum-codigos
  /vade-mecum/codigos/import/estatuto:
    post:
      consumes:
      - multipart/form-data
      description: Importa registros de estatuto utilizando um arquivo Excel (.xlsx)
        com cabeçalho padrão
      parameters:
      - description: Arquivo Excel (.xlsx)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Importar estatutos via Excel
      tags:
      - vade-mecum-codigos
...
securityDefinitions:
  Bearer:
    description: Digite "Bearer" seguido do token JWT
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"